requirements.txt
unicorn-engine>=2.0.0
capstone>=5.0.0
pefile>=2023.2.7
